Move CheckStructFields to a separate package
